Rename file to `token_selector.py`
