+++ shortfin/tests/apps/llm/components/token_selection_strategy/independent_token_selection_strategy_test.py
@@ -29,7 +29,7 @@ from shortfin_apps.llm.components.token_selection_strategy import (
     TokenSelector,
     TokenSelectionStrategy,
 )
-from shortfin_apps.llm.components.token_selection_strategy.independent_token_selection_strategy import (
+from shortfin_apps.llm.components.token_selection_strategy.token_selector import (
     Beam,
 )
